requirements.txt
flask
requests
beautifulsoup4
gunicorn
python-dotenv
pip install flask-cors
